refactor(core): decompose database configuration
- Extract validation module
- Extract connection listeners
- Extract health probes
- Preserve database bootstrap API and reduce db.ts from 581 to 294 LOC
